install: Optimise the distro package installation
Given that the role's minimum Ansible version is set
to 2.2 we can assume that the package module is an
action module and the underlying modules are able to
take a package list for the 'name' argument'.

As such, we can pass the package list to the task
which makes the installation slightly more efficient.

Note: This does not work for Ansible 2.1 and earlier.