install: Fix cache update after initial apt_repository fail
Given that updating the apt cache/keys is a network operation,
they are prone to transient failures and should therefore
include the use of retries on the task.

However, as noted in [1], if apt_repository fails to update
the apt cache after updating the configuration, retries do
not register there was a change and so no attempt is made
to update the cache by the module on the second attempt.
This failure can result in a failure to install packages.

This change adds an apt module task to update the cache if
the apt_repository tasks or the apt_key tasks register a
change. This means updating the cache will get retried on
failure and no longer fail silently.

The apt cache update is triggered on changes to the apt
repository configuration and to the keys to ensure that the
cache is always using the right config/key conbination.

[1] https://github.com/ansible/ansible/issues/36605