install: Use a ternary filter for 'present' vs 'latest'
It is not necessary to use two tasks with a conditional on
each to achieve the goal of installing using present/lastest
on the package task. Instead we use a ternary filter for the
state argument. This achieves the same goal and is slightly
more efficient in terms of both code and execution.