vars: Handle the use of grafana_version='latest' properly
When grafana_version='latest', the package handlers do not
know what to do with the package name given to them, and
the task fails.

This patch implements a ternary check in the vars to ensure
that if that value is given, the version is omitted from
the package name given to the task which installs it.