datasources: Allow additional datasources to be created later
Currently the role only allows the datasources to be configured
on the first implementation. From then on if there are any data
sources configured, the task to upload any new configuration is
ignored.

This patch implements a validation for each datasource item to
check whether it's already configured using the name as the key.
If the item is not configured, it will configure it.