Add ability to generate API keys
This patch adds the ability to generate API keys
which can be used in automation later in the
same playbook, or by entirely different playbooks.
Given that the API keys are generated once-off
and not retrievable after the fact, we store the
keys on the deployment host to allow retrieval
and re-use.