Make template compatible with both Python 2 and 3
When using this role in an environment where the default python
interpreter is python3, the ansible run fails when jinja2 tries to
process `templates/grafana.ini.j2`.

This is because in Python 3, `dict` objects no longer have the
`.iteritems()` method (see e.g.: https://wiki.python.org/moin/Python3.0,
under "Built-in Changes").

For the purposes it's used in the template, `.items()` is functionally
equivalent to `.iteritems()` except that it also works in Python 3.