Update datasource regex match to include dashes and numbers
In order to match datasources like "DS_DML-PROM0" which include dashes
and numbers, we open the match to include them. We also remove the
capture groups implemented in [1] as they're not required and their
inclusion makes things more difficult to understand. Instead, we provide
the same function by simply also allowing underscores in the match.

[1] https://github.com/cloudalchemy/ansible-grafana/pull/53

Fixes #59