Correct package source for ARM in README. (#149)
Since #83 began using the new official packages for ARM architectures the old README text was out of correct. I've retained references to the history (so as not to confuse anyone who might be using older versions of this project) but also updated the description of the current source of packages.