Cleaner and more general handling of auth options
Before this change, the keys of grafana_auth's value were tested for a
static list of names like 'disable_login_form'. If they didn't match it
would suppose they're a mapping and thus descent into it.
This would work as long as the key was in the list, but for example
using 'login_maximum_inactive_lifetime_days' would fail as it's not a
hash, but it is a valid grafana option.

This change makes the test more elegant, robust and easier to maintain.

Fixes #218